+++ src/Headless.Payments.Paymob.CashIn/Models/Orders/CashInCreateOrderRequest.cs
@@ -18,10 +18,13 @@ public sealed class CashInCreateOrderRequest
 {
     private CashInCreateOrderRequest() { }
 
-    public int AmountCents { get; private init; }
+    /// <summary>The order amount in the smallest currency unit (integer cents), e.g. <c>10000</c> for 100.00 EGP.</summary>
+    public long AmountCents { get; private init; }
 
+    /// <summary>ISO 4217 currency code for the order (for example <c>EGP</c>).</summary>
     public string Currency { get; private init; } = null!;
 
+    /// <summary>Your own order reference ID, used to correlate Paymob orders with your system. Optional.</summary>
     public string? MerchantOrderId { get; private init; }
 
     /// <summary>
@@ -52,7 +55,7 @@ public sealed class CashInCreateOrderRequest
     /// <exception cref="ArgumentOutOfRangeException"><paramref name="amountCents"/> is not positive.</exception>
     /// <exception cref="ArgumentException"><paramref name="currency"/> is null or empty.</exception>
     public static CashInCreateOrderRequest CreateOrder(
-        int amountCents,
+        long amountCents,
         string currency = "EGP",
         string? merchantOrderId = null
     )
@@ -84,7 +87,7 @@ public sealed class CashInCreateOrderRequest
         CashInCreateOrderRequestShippingDetails shippingDetails,
         CashInCreateOrderRequestShippingData shippingData,
         ICollection<CashInCreateOrderRequestOrderItem> items,
-        int amountCents,
+        long amountCents,
         string currency = "EGP",
         string? merchantOrderId = null
     )
