+++ src/Headless.Payments.Paymob.CashIn/Models/Payment/CashInPaymentKeyRequest.cs
@@ -7,10 +7,10 @@ namespace Headless.Payments.Paymob.CashIn.Models.Payment;
 public sealed class CashInPaymentKeyRequest
 {
     public CashInPaymentKeyRequest(
-        int integrationId,
-        int orderId,
+        long integrationId,
+        long orderId,
         CashInBillingData billingData,
-        int amountCents,
+        long amountCents,
         string currency = "EGP",
         bool lockOrderWhenPaid = true,
         int? expiration = null
@@ -36,11 +36,12 @@ public sealed class CashInPaymentKeyRequest
         BillingData = billingData;
     }
 
-    public int IntegrationId { get; }
+    public long IntegrationId { get; }
 
-    public int OrderId { get; }
+    public long OrderId { get; }
 
-    public int AmountCents { get; }
+    /// <summary>The amount in the smallest currency unit (integer cents), e.g. <c>10000</c> for 100.00 EGP.</summary>
+    public long AmountCents { get; }
 
     public string Currency { get; }
